pyproject.toml
[tool.poetry]
name = "anchor_platform_tutorial"
version = "0.0.1"
description = ""
authors = ["Jake Urban <jake@stellar.org>"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "*"
requests = "*"
flask-sqlalchemy = "^3.0.2"
python-dotenv = "^0.21.0"
Flask-Migrate = "^4.0.0"
PyJWT = "^2.6.0"

[tool.poetry.dev-dependencies]
black = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
